feat: logic
